unit tests

+++ jardin/cache_stores/disk.py
@@ -1,6 +1,7 @@
 import os
 import pyarrow.feather as feather
 import pandas as pd
+import time
 from datetime import datetime
 
 from jardin.cache_stores.base import Base
@@ -9,9 +10,17 @@ class Disk(Base):
     
     EXTENSION = '.feather'
 
-    def __init__(self, dir, limit=None):
+    def __init__(self, dir, limit=None, ttl=None):
+        
+        """
+            dir: path of directory to use for cached files
+            limit: size limit in bytes for lru cache
+            ttl: cache expiry in seconds
+        """
+        
         self.dir = dir
         self.limit = limit
+        self.ttl = ttl
         if not os.path.exists(self.dir):
             try:
                 os.makedirs(self.dir)
@@ -26,19 +35,26 @@ class Disk(Base):
     def __setitem__(self, key, value):
         if not isinstance(value, pd.DataFrame):
             return None
+        
         feather.write_feather(value, self._path(key))
 
         if self.limit is not None:
             if os.path.getsize(self._path(key)) > self.limit:
-                raise MemoryError(f"disk cache limit exceeded by key {key}")
+                raise MemoryError(f"disk cache limit exceeded by single key {key}")
             while self.size() > self.limit:
                 del self[self.lru()]
 
     def __delitem__(self, key):
-        os.remove(self._path(key))
+        if os.path.exists(self._path(key)):
+            os.remove(self._path(key))
 
     def __contains__(self, key):
-        return os.path.isfile(self._path(key))
+        f = self._path(key)
+        if not os.path.isfile(f):
+            return False
+        if self.ttl is None:
+            return True
+        return int(time.time() - os.stat(f).st_mtime) < self.ttl
 
     def __len__(self):
         return len(self.keys())
